+++ packages/stacked_services/example/lib/app/routes.gr.dart
@@ -7,7 +7,6 @@
 // ignore_for_file: public_member_api_docs
 
 import 'package:auto_route/auto_route.dart';
-import 'package:auto_route/legacy.dart';
 import 'package:flutter/material.dart';
 
 import '../ui/views/first_screen.dart';